Move loopback-host tests into their own suite; document CmxManualPairingEntry.init
